pyproject.toml
[project]
name = "nfl-analytics"
version = "0.1.0"
description = "NFL Sports Betting Analytics & Research Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "meteostat>=1.6.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "plotly>=5.22.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "ipython>=8.26.0",
    "jupyter>=1.0.0",
    "pre-commit>=3.8.0",
]
torch = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
]
research = [
    "jupyterlab>=4.2.0",
    "ipywidgets>=8.1.0",
    "nbformat>=5.10.0",
    "quarto>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=py --cov-report=html --cov-report=term"